% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{check_km_curve}
\alias{check_km_curve}
\title{Check KM curve for validity}
\usage{
check_km_curve(S.KM, group_name = "Group", stop_on_error = TRUE)
}
\arguments{
\item{S.KM}{Survival probabilities}

\item{group_name}{Name of the group}

\item{stop_on_error}{Whether to stop on error (default TRUE)}
}
\description{
Check KM curve for validity
}
