% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{count_weighted}
\alias{count_weighted}
\title{Weighted counting process}
\usage{
count_weighted(x, y, w = rep(1, length(y)))
}
\arguments{
\item{x}{Time point.}

\item{y}{Vector of event/censoring times.}

\item{w}{Weights (default 1).}
}
\value{
Weighted count of events up to time x.
}
\description{
Computes the weighted count of events up to a specified time.
}
