% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedCox.R
\name{cox_rhogamma_resample}
\alias{cox_rhogamma_resample}
\title{Resampling for Weighted Cox Model (rho, gamma)}
\usage{
cox_rhogamma_resample(
  fit_rhogamma,
  i_bhat,
  K_wt_rg,
  i_zero,
  K_zero,
  G1.draws = NULL,
  G0.draws = NULL,
  draws = 100,
  seedstart = 8316951,
  workers = NULL
)
}
\arguments{
\item{G1.draws, }{G0.draws Optional: pre-generated random draws for groups}

\item{draws}{Number of resampling iterations}

\item{workers}{Number of parallel workers}

\item{bhat}{Estimated coefficient}

\item{time}{Event/censoring times}

\item{delta}{Event indicator}

\item{z}{Treatment group indicator (0/1)}

\item{w_hat}{Subjects' weighting (eg. propensity-scores)}

\item{rho, }{gamma Weighting parameters}

\item{t.tau, }{w0.tau, w1.tau Optional: custom weights}

\item{seed.value}{Optional: random seed}

\item{parallel}{Logical: use parallelization? (default FALSE)}
}
\value{
List with resampling results
}
\description{
Resampling for Weighted Cox Model (rho, gamma)
}
