% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting_helpers.R
\name{get_dfcounting}
\alias{get_dfcounting}
\title{Get df_counting}
\usage{
get_dfcounting(
  df,
  tte.name,
  event.name,
  treat.name,
  arms,
  by.risk = 12,
  cox.digits = 3,
  lr.digits = 3,
  qprob = 0.5,
  strata.name = NULL,
  weight.name = NULL,
  check.KM = TRUE,
  scheme = "fh",
  scheme_params = list(rho = 0, gamma = 0),
  draws = 0,
  seedstart = 8316951,
  check.seKM = FALSE
)
}
\description{
Get df_counting
}
