% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting_helpers.R
\name{get_riskpoints}
\alias{get_riskpoints}
\title{Get risk set counts at specified risk points}
\usage{
get_riskpoints(ybar, risk_points, at_points)
}
\arguments{
\item{ybar}{Numeric vector of risk set counts.}

\item{risk_points}{Numeric vector of risk points.}

\item{at_points}{Numeric vector of time points.}
}
\value{
Numeric vector of risk set counts at risk points.
}
\description{
Returns risk set counts at specified risk points.
}
