% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{kmq_calculations}
\alias{kmq_calculations}
\title{Kaplan-Meier quantile calculation}
\usage{
kmq_calculations(time_points, survival_probs, qprob = 0.5, type = "midpoint")
}
\arguments{
\item{time_points}{Vector of time points.}

\item{survival_probs}{Vector of survival probabilities.}

\item{qprob}{Quantile probability (default 0.5).}

\item{type}{Calculation type (midpoint or min).}
}
\value{
Estimated quantile time.
}
\description{
Kaplan-Meier quantile calculation
}
