% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{resampling_survival}
\alias{resampling_survival}
\title{Resampling Survival Curves for Confidence Bands}
\usage{
resampling_survival(U, W, D, at.points, draws.band, surv, G_draws)
}
\arguments{
\item{U}{Vector of observed times.}

\item{W}{Vector of weights.}

\item{D}{Vector of event indicators (0/1).}

\item{at.points}{Vector of time points for evaluation.}

\item{draws.band}{Number of resampling draws.}

\item{surv}{Vector of survival estimates.}

\item{G_draws}{Matrix of random draws for resampling.}
}
\value{
Matrix of resampled survival curves.
}
\description{
Performs resampling to generate survival curves for a group, used for constructing confidence bands.
}
