% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{risk_weighted}
\alias{risk_weighted}
\title{Weighted risk set}
\usage{
risk_weighted(x, y, w = rep(1, length(y)))
}
\arguments{
\item{x}{Time point.}

\item{y}{Vector of event/censoring times.}

\item{w}{Weights (default 1).}
}
\value{
Weighted number at risk at time x.
}
\description{
Weighted risk set
}
