% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedCox.R
\name{score_calculation}
\alias{score_calculation}
\title{Score calculation for weighted Cox model}
\usage{
score_calculation(ybar1, ybar0, dN1, dN0, wt_rg)
}
\arguments{
\item{ybar1}{Numeric vector of event counts for group 1.}

\item{ybar0}{Numeric vector of event counts for group 0.}

\item{dN1}{Numeric vector of event increments for group 1.}

\item{dN0}{Numeric vector of event increments for group 0.}

\item{wt_rg}{Numeric vector of rho-gamma weights.}
}
\value{
List with score, variance, information, and weights.
}
\description{
Calculates the score and variance for the weighted Cox model.
}
