% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_helpers.R
\name{wt.rg.S}
\alias{wt.rg.S}
\title{Compute Time-Dependent Weights for Survival Analysis}
\usage{
wt.rg.S(
  S,
  scheme = c("fh", "schemper", "XO", "MB", "custom_time", "fh_exp1", "fh_exp2"),
  rho = NULL,
  gamma = NULL,
  Scensor = NULL,
  Ybar = NULL,
  tpoints = NULL,
  t.tau = NULL,
  w0.tau = 0,
  w1.tau = 1,
  mb_tstar = NULL,
  details = FALSE
)
}
\description{
Compute Time-Dependent Weights for Survival Analysis
}
