% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{add_legends}
\alias{add_legends}
\title{Add legends to KM plot}
\usage{
add_legends(
  dfcount,
  show.cox,
  cox.cex,
  put.legend.cox,
  show.logrank,
  logrank.cex,
  put.legend.lr,
  show_arm_legend,
  arms,
  col.1,
  col.0,
  ltys,
  lwds,
  arm.cex,
  put.legend.arms
)
}
\arguments{
\item{dfcount}{List with results}

\item{show.cox}{Show Cox legend}

\item{cox.cex}{Cox legend size}

\item{put.legend.cox}{Cox legend position}

\item{show.logrank}{Show logrank legend}

\item{logrank.cex}{Logrank legend size}

\item{put.legend.lr}{Logrank legend position}

\item{show_arm_legend}{Show arm legend}

\item{arms}{Arm labels}

\item{col.1}{Color for group 1}

\item{col.0}{Color for group 0}

\item{ltys}{Line types}

\item{lwds}{Line widths}

\item{arm.cex}{Arm legend size}

\item{put.legend.arms}{Arm legend position}
}
\description{
Add legends to KM plot
}
