% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{add_median_annotation}
\alias{add_median_annotation}
\title{Add median annotation to KM plot}
\usage{
add_median_annotation(
  medians_df,
  med.digits,
  med.cex,
  med.font,
  xmed.fraction,
  ymed.offset
)
}
\arguments{
\item{medians_df}{Data frame with quantile results.}

\item{med.digits}{Integer; digits for median.}

\item{med.cex}{Numeric; text size for median.}

\item{med.font}{Integer; font for median.}

\item{xmed.fraction}{Numeric; x position fraction.}

\item{ymed.offset}{Numeric; y offset.}
}
\description{
Adds median survival annotation to a Kaplan-Meier plot.
}
