% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{add_median_annotation}
\alias{add_median_annotation}
\title{Add median annotation to KM plot}
\usage{
add_median_annotation(
  medians_df,
  med.digits,
  med.cex,
  med.font,
  xmed.fraction,
  ymed.offset
)
}
\arguments{
\item{medians_df}{Data frame with quantile results}

\item{med.digits}{Digits for median}

\item{med.cex}{Text size for median}

\item{med.font}{Font for median}

\item{xmed.fraction}{X position fraction}

\item{ymed.offset}{Y offset}
}
\description{
Add median annotation to KM plot
}
