% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{add_risk_table}
\alias{add_risk_table}
\title{Add risk table annotation to KM plot}
\usage{
add_risk_table(
  risk.points,
  rpoints0,
  rpoints1,
  col.0,
  col.1,
  risk.cex,
  ymin,
  risk_offset,
  risk_delta,
  y.risk0,
  y.risk1
)
}
\arguments{
\item{risk.points}{Risk time points}

\item{rpoints0}{Risk set for group 0}

\item{rpoints1}{Risk set for group 1}

\item{col.0}{Color for group 0}

\item{col.1}{Color for group 1}

\item{risk.cex}{Text size for risk table}

\item{ymin}{Minimum y value}

\item{risk_offset}{Offset for risk table}

\item{risk_delta}{Delta for risk table}

\item{y.risk0}{Y position for group 0 risk table}

\item{y.risk1}{Y position for group 1 risk table}
}
\description{
Add risk table annotation to KM plot
}
