% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting_helpers.R
\name{check_results}
\alias{check_results}
\title{Checking results}
\usage{
check_results(dfcount)
}
\arguments{
\item{dfcount}{Result object from df_counting.}
}
\value{
None. Prints summary statistics.
}
\description{
Prints summary statistics for logrank and Cox model results from dfcounting.
}
