% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting_helpers.R
\name{get_dfcounting}
\alias{get_dfcounting}
\title{Get df_counting results}
\usage{
get_dfcounting(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{df_counting}}.}
}
\value{
Result from df_counting or NULL if error.
}
\description{
Wrapper for df_counting, safely executes and returns results for survival analysis.
}
\seealso{
\code{\link{df_counting}}
}
