% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{get_event_risk_matrices}
\alias{get_event_risk_matrices}
\title{Event and Risk Matrices for Survival Analysis}
\usage{
get_event_risk_matrices(U, at.points)
}
\arguments{
\item{U}{Vector of observed times (e.g., time-to-event).}

\item{at.points}{Vector of time points at which to evaluate events and risk.}
}
\value{
A list with elements:
\item{event_mat}{Matrix indicating if event occurred by each time point.}
\item{risk_mat}{Matrix indicating if subject is at risk at each time point.}
}
\description{
Constructs matrices indicating event and risk status for each subject at specified time points.
}
\examples{
U <- c(1, 2, 3)
at.points <- c(1, 2, 3)
get_event_risk_matrices(U, at.points)

}
