% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_wlr_calculations_helpers.R
\name{KM_diff}
\alias{KM_diff}
\title{KM difference between groups}
\usage{
KM_diff(
  df,
  tte.name = "tte",
  event.name = "event",
  treat.name = "treat",
  weight.name = NULL,
  at_points = sort(df[[tte.name]]),
  alpha = 0.05,
  seedstart = 8316951,
  draws = 0,
  risk.points,
  draws.band = 0,
  tau.seq = 0.25,
  qtau = 0.025,
  show_resamples = TRUE,
  modify_tau = FALSE
)
}
\arguments{
\item{df}{Data frame with survival data.}

\item{tte.name}{Name of time-to-event variable.}

\item{event.name}{Name of event indicator variable.}

\item{treat.name}{Name of treatment group variable.}

\item{weight.name}{Name of weights variable.}

\item{at_points}{Time points for calculation.}

\item{alpha}{Significance level.}

\item{seedstart}{Random seed.}

\item{draws}{Number of resampling draws.}
}
\value{
Data frame with KM differences.
}
\description{
Calculates the difference in KM curves between two groups.
}
