% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{plot_km_confint_polygon}
\alias{plot_km_confint_polygon}
\title{Plot confidence interval polygon for KM curve}
\usage{
plot_km_confint_polygon(x, surv, se, conf_level, col)
}
\arguments{
\item{x}{Time points}

\item{surv}{Survival probabilities}

\item{se}{Standard errors of survival probabilities}

\item{conf_level}{Confidence level (default 0.95)}

\item{col}{Polygon color}
}
\description{
Plot confidence interval polygon for KM curve
}
