% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{weightedSurv-package}
\alias{weightedSurv}
\alias{weightedSurv-package}
\title{weightedSurv: Survival Analysis with Subject-Specific (Case Weights) and Time-Dependent Weighting}
\description{
Survival analysis functions allowing for time-dependent and subject-specific (for example, propensity-scores) weighting. Weighted estimation for Cox model, Kaplan-Meier treatment survival curves as well as treatment difference along with point-wise and simultaneous confidence bands, and restricted mean survival time comparisons where 'RMST' estimates are evaluated across all potential truncation times (point-wise and simultaneous bands).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/larry-leon/weightedSurv}
  \item Report bugs at \url{https://github.com/larry-leon/weightedSurv/issues}
}

}
\author{
\strong{Maintainer}: Larry Leon \email{larry.leon.05@post.harvard.edu}

}
\keyword{internal}
